rename pkg database, commands, handlers
