+++ main.go
@@ -7,8 +7,8 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/nurali-techie/rc/cfg"
 	"github.com/nurali-techie/rc/cli"
-	"github.com/nurali-techie/rc/commands"
-	"github.com/nurali-techie/rc/database"
+	"github.com/nurali-techie/rc/command"
+	"github.com/nurali-techie/rc/db"
 	"github.com/nurali-techie/rc/io"
 	"github.com/nurali-techie/rc/service"
 	"github.com/nurali-techie/rc/store"
@@ -21,7 +21,7 @@ func main() {
 	cfg.Init()
 
 	// setup database
-	db, closeFn := database.GetDatabase()
+	db, closeFn := db.GetDatabase()
 	defer closeFn(db)
 
 	// setup dependency
@@ -33,14 +33,14 @@ func main() {
 	console := io.NewConsole()
 
 	// init commander
-	helpCmd := commands.NewHelpCommand(console)
-	getCmd := commands.NewGetCommand(commandService, clipboard)
+	helpCmd := command.NewHelpCommand(console)
+	getCmd := command.NewGetCommand(commandService, clipboard)
 	commander := cli.NewCommander(helpCmd, getCmd)
 
 	// register commands
-	commander.Register("add", commands.NewAddCommand(commandService, clipboard))
-	commander.Register("ls", commands.NewListCommand(commandService, console))
-	commander.Register("web", commands.NewWebCommand(commandService))
+	commander.Register("add", command.NewAddCommand(commandService, clipboard))
+	commander.Register("ls", command.NewListCommand(commandService, console))
+	commander.Register("web", command.NewWebCommand(commandService))
 
 	// execute command
 	err = commander.ServeCommand(os.Args[1:])
